backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
firebase-admin==6.6.0
httpx==0.27.2
pydantic==2.10.3
python-dotenv==1.0.1
livekit==0.17.3
livekit-agents==0.9.6
livekit-plugins-google==0.8.3
livekit-api==0.7.1
